+++ hT4_3/src/Expression3.java
@@ -1,4 +1,4 @@
 public interface Expression3 {
-    public double evaluate(double var1, double var2, double var3);
-    //public int evaluate(int var1,int var2,int var3);
+    //public double evaluate(double var1, double var2, double var3);
+    public int evaluate(int var1,int var2,int var3);
 }
